+++ src/server/life/ChangeableStats.java
@@ -45,13 +45,13 @@ public class ChangeableStats extends OverrideMonsterStats {
         final double mod = (double) newLevel / (double) stats.getLevel();
         final double hpRatio = (double) stats.getHp() / (double) stats.getExp();
         final double pqMod = (pqMob ? 1.5 : 1.0); // god damn
-        hp = (long) Math.round((!stats.isBoss() ? GameConstants.getMonsterHP(newLevel) : (stats.getHp() * mod)) * pqMod); // right here lol
+        hp = Math.round((!stats.isBoss() ? GameConstants.getMonsterHP(newLevel) : (stats.getHp() * mod)) * pqMod); // right here lol
         exp = (int) Math.round((!stats.isBoss() ? (GameConstants.getMonsterHP(newLevel) / hpRatio) : (stats.getExp())));
         mp = (int) Math.round(stats.getMp() * mod * pqMod);
         watk = (int) Math.round(stats.getPhysicalAttack() * mod);
         matk = (int) Math.round(stats.getMagicAttack() * mod);
-        acc = (int) Math.round(stats.getAcc() + Math.max(0, newLevel - stats.getLevel()) * 2);
-        eva = (int) Math.round(stats.getEva() + Math.max(0, newLevel - stats.getLevel()));
+        acc = Math.round(stats.getAcc() + Math.max(0, newLevel - stats.getLevel()) * 2);
+        eva = Math.round(stats.getEva() + Math.max(0, newLevel - stats.getLevel()));
         PDRate = Math.min(stats.isBoss() ? 30 : 20, (int) Math.round(stats.getPDRate() * mod));
         MDRate = Math.min(stats.isBoss() ? 30 : 20, (int) Math.round(stats.getMDRate() * mod));
         pushed = (int) Math.round(stats.getPushed() * mod);
